Optimized the code. Added Options and Settings functionality. Implemented NSFW detection. Adjusted cache directory structure. Made some UI interface adjustments.

+++ server.py
@@ -1,7 +1,9 @@
 print("starting...",flush=True)
 
+import subprocess
 import sys
 import shutil
+import time
 import glob
 from http.server import BaseHTTPRequestHandler
 from http.server import ThreadingHTTPServer
@@ -12,11 +14,13 @@ import roop.core
 import roop.face_analyser
 from roop.processors.frame import face_swapper
 import roop.globals
+import roop.predictor
 import threading
 import cv2
 import numpy as np
 from roop import utilities
 from tqdm import tqdm
+import onnxruntime
 
 enhancer=None
 def face_enhance(target_face,frame_img):
@@ -27,13 +31,24 @@ def face_enhance(target_face,frame_img):
         enhancer.pre_check()
     return enhancer.enhance_face(target_face,frame_img)
 
+def run_ffmpeg(args) -> bool:
+    commands = ['ffmpeg', '-hide_banner', '-loglevel', roop.globals.log_level]
+    commands.extend(args)
+    print(' '.join(commands))
+    try:
+        subprocess.check_output(commands)
+        return True
+    except Exception:
+        pass
+    return False
+
 def extract_frames(input_path: str,output_dir:str, fps: float = 30) -> bool:
     temp_frame_quality = roop.globals.temp_frame_quality * 31 // 100
-    utilities.run_ffmpeg(['-hwaccel', 'auto', '-i', input_path, '-q:v', str(temp_frame_quality), '-pix_fmt', 'rgb24', '-vf', 'fps=' + str(fps), os.path.join(output_dir, '%04d.' + roop.globals.temp_frame_format)])
+    run_ffmpeg(['-hwaccel', 'auto', '-i', input_path, '-q:v', str(temp_frame_quality), '-pix_fmt', 'rgb24', '-vf', 'fps=' + str(fps), os.path.join(output_dir, '%04d.' + roop.globals.temp_frame_format)])
 
 def create_gif(input_frames_dir,fps, output_path):
-    input_frames_dir=os.path.join(input_frames_dir,"%04d.png")
-    return utilities.run_ffmpeg(['-hwaccel', 'auto', '-f', 'image2', '-r', str(fps), '-i', input_frames_dir,output_path])
+    input_frames_dir=os.path.join(input_frames_dir,"%04d."+roop.globals.temp_frame_format)
+    return run_ffmpeg(['-hwaccel', 'auto', '-f', 'image2', '-r', str(fps), '-i', input_frames_dir,output_path])
 
 def create_video(input_frames_dir, fps, output_path):
     output_dir = os.path.join(input_frames_dir, '%04d.' + roop.globals.temp_frame_format)
@@ -48,10 +63,10 @@ def create_video(input_frames_dir, fps, output_path):
     if roop.globals.output_video_encoder in ['h264_nvenc', 'hevc_nvenc']:
         commands.extend(['-cq', str(output_video_quality)])
     commands.extend(['-pix_fmt', 'yuv420p', '-vf', 'colorspace=bt709:iall=bt601-6-625:fast=1', '-y', output_path])
-    return utilities.run_ffmpeg(commands)
+    return run_ffmpeg(commands)
 
 def add_audio(video_path:str,audio_path: str, output_path: str) -> bool:
-    return utilities.run_ffmpeg(['-hwaccel', 'auto', 
+    return run_ffmpeg(['-hwaccel', 'auto', 
                                  '-i', video_path, 
                                  '-i', audio_path, 
                                  '-c:v', 'copy', 
@@ -78,6 +93,14 @@ def get_most_similar_face(faces, face_to_find,min_similarity):
     if max_similarity>=min_similarity:
         return max_similarityFace
     return None
+    
+def clean_temp(dir):
+    if not roop.globals.keep_frames:
+        if os.path.exists(dir):
+            shutil.rmtree(dir)
+    parent_dir=os.path.dirname(dir)
+    if os.path.exists(parent_dir) and not os.listdir(parent_dir):
+        os.rmdir(parent_dir)
 
 class JsonCustomEncoder(json.JSONEncoder): 
     def default(self, field): 
@@ -90,16 +113,24 @@ class JsonCustomEncoder(json.JSONEncoder):
         else:
             return json.JSONEncoder.default(self, field)
 
-def cv2imread(file_path):
-    cv_img=cv2.imdecode(np.fromfile(file_path,dtype=np.uint8),cv2.IMREAD_COLOR)
-    return cv_img
+def cv2imread2rgb(file_path):
+    #return cv2.imread(file_path)
+    return cv2.imdecode(np.fromfile(file_path,dtype=np.uint8),cv2.IMREAD_COLOR)
+
+def rgb2bgr(imgdata):
+    #return imgdata
+    return cv2.cvtColor(imgdata,cv2.COLOR_RGB2BGR)
+
+def cv2imwrite(output_path,imgdata):
+    #return cv2.imwrite(output_path,imgdata)
+    return cv2.imencode("."+os.path.splitext(output_path)[-1],imgdata)[1].tofile(output_path)
 
 def process_frames(source_face_infos,target_face_infos,frame_paths,output_dir, progress,min_similarity):
     for frame_path in frame_paths:
-        frame = cv2imread(frame_path)
+        frame = cv2imread2rgb(frame_path)
         try:
             result = process_one_frame(source_face_infos,target_face_infos, frame,min_similarity)
-            cv2.imwrite(os.path.join(output_dir,os.path.split(frame_path)[-1]), result)
+            cv2imwrite(os.path.join(output_dir,os.path.split(frame_path)[-1]), result)
         except Exception:
             print(f"process frame err:{traceback.format_exc()}")
             pass
@@ -107,7 +138,7 @@ def process_frames(source_face_infos,target_face_infos,frame_paths,output_dir, p
             progress.update(1)
 
 def process_one_frame(source_face_infos,target_face_infos,frame,min_similarity):
-    frame_faces=roop.face_analyser.get_many_faces(cv2.cvtColor(frame,cv2.COLOR_RGB2BGR))
+    frame_faces=roop.face_analyser.get_many_faces(rgb2bgr(frame))
     result=frame
     if frame_faces is not None and len(frame_faces)>0:
         min_len=min(len(source_face_infos),len(target_face_infos))
@@ -185,29 +216,36 @@ class MyHTTPHandler(BaseHTTPRequestHandler):
         self.wfile.write(resultStr.encode('utf-8'))
     
     def func_get_faces(self,input_img_path):
-        img=cv2imread(input_img_path)
-        t=roop.face_analyser.get_many_faces(cv2.cvtColor(img,cv2.COLOR_RGB2BGR))
+        img=cv2imread2rgb(input_img_path)
+        t=roop.face_analyser.get_many_faces(rgb2bgr(img))
         faces=[]
         if t is not None:
             for item in t:
                 faces.append(item['bbox'])
         return {'width':img.shape[1],'height':img.shape[0],'faces':faces}
     
-    def func_swap_video(self,source_face_infos,target_face_infos,target_path,output_path,keep_fps,min_similarity):
+    def func_swap_video(self,source_face_infos,target_face_infos,target_path,output_path,min_similarity):
+        print("Checking nsfw...")
+        if roop.predictor.predict_video(target_path):
+            return 'nsfw'
         video_name_full = os.path.split(target_path)[-1]
-        video_name = os.path.splitext(video_name_full)[0]
-        temp_output_dir = os.path.join(os.path.dirname(target_path),"faceswap_temp",video_name)
+        temp_output_dir_name=video_name_full.replace('.','_')+"_"+os.path.splitext(os.path.split(output_path)[-1])[0]
+        temp_output_dir = os.path.join(os.path.dirname(target_path),"faceswap_temp",temp_output_dir_name)
         if os.path.exists(temp_output_dir):
             shutil.rmtree(temp_output_dir)
         os.makedirs(temp_output_dir,exist_ok=True)
         print("detecting video's FPS...")
         fps = utilities.detect_fps(target_path)
-        if not keep_fps:
+        if not roop.globals.keep_fps:
             fps=30
         print(f'Extracting frames with {fps} FPS...')
-        extract_frames(target_path,temp_output_dir,fps)
+        original_frames_temp_dir=os.path.join(temp_output_dir,"original")
+        if os.path.exists(original_frames_temp_dir):
+            shutil.rmtree(original_frames_temp_dir)
+        os.makedirs(original_frames_temp_dir,exist_ok=True)
+        extract_frames(target_path,original_frames_temp_dir,fps)
         frame_paths = tuple(sorted(
-            glob.glob(os.path.join(temp_output_dir,"*.png")),
+            glob.glob(os.path.join(original_frames_temp_dir,"*."+roop.globals.temp_frame_format)),
             key=lambda x: int(os.path.splitext(os.path.split(x)[-1])[0])
         ))
         
@@ -217,7 +255,7 @@ class MyHTTPHandler(BaseHTTPRequestHandler):
             if info["file"] in source_cache:
                 source_img_faces=source_cache[info["file"]]
             else:
-                source_img=cv2.cvtColor(cv2imread(info["file"]),cv2.COLOR_RGB2BGR)
+                source_img=rgb2bgr(cv2imread2rgb(info["file"]))
                 source_img_faces=roop.face_analyser.get_many_faces(source_img)
                 source_cache[info["file"]]=source_img_faces
             info["face"]=source_img_faces[info["face_index"]]
@@ -227,38 +265,83 @@ class MyHTTPHandler(BaseHTTPRequestHandler):
             if info["file"] in target_cache:
                 target_img_faces=target_cache[info["file"]]
             else:
-                target_img=cv2.cvtColor(cv2imread(info["file"]),cv2.COLOR_RGB2BGR)
+                target_img=rgb2bgr(cv2imread2rgb(info["file"]))
                 target_img_faces=roop.face_analyser.get_many_faces(target_img)
                 target_cache[info["file"]]=target_img_faces
             info["face"]=target_img_faces[info["face_index"]]
-        frames_temp_dir = os.path.join(os.path.dirname(temp_output_dir),video_name+"_swapped")
-        if os.path.exists(frames_temp_dir):
-            shutil.rmtree(frames_temp_dir)
+        frames_temp_dir = os.path.join(temp_output_dir,"swapped")
         os.makedirs(frames_temp_dir,exist_ok=True)
         process_video(source_face_infos,target_face_infos,frame_paths,frames_temp_dir,min_similarity)
         if video_name_full.endswith(".gif"):
             print("creating gif...")
             if not create_gif(frames_temp_dir,fps,output_path):
-                return 'fail'
+                clean_temp(temp_output_dir)
+                return 'Create gif failed.'
         else:
             print("creating video...")
-            temp_output_path=os.path.join(frames_temp_dir,"tempvideo.mp4")
-            if not create_video(frames_temp_dir, fps, temp_output_path):
-                return 'fail'
-            print("adding audio...")
-            if not add_audio(temp_output_path, target_path, output_path):
-                print("add audio failed")
-        return 'succ'
+            if roop.globals.skip_audio:
+                if not create_video(frames_temp_dir, fps, output_path):
+                    clean_temp(temp_output_dir)
+                    return 'Create video failed.'
+            else:
+                temp_output_path=os.path.join(temp_output_dir,"tempvideo.mp4")
+                if not create_video(frames_temp_dir, fps, temp_output_path):
+                    if os.path.exists(temp_output_path):
+                        os.remove(temp_output_path)
+                    clean_temp(temp_output_dir)
+                    return 'Create video failed.'
+                print("adding audio...")
+                if not add_audio(temp_output_path, target_path, output_path):
+                    print("add audio failed")
+                if os.path.exists(temp_output_path):
+                    os.remove(temp_output_path)
+        clean_temp(temp_output_dir)
+        return 'succ'        
     
     def func_video_screenshot(self,duration,input_path,output_path):
-        if utilities.run_ffmpeg(['-hwaccel','auto', '-ss' ,str(duration),'-i',input_path,'-r','1','-vframes','1','-an','-vcodec','mjpeg','-y',output_path]):
+        if run_ffmpeg(['-hwaccel','auto', '-ss' ,str(duration),'-i',input_path,'-r','1','-vframes','1','-an','-vcodec','mjpeg','-y',output_path]):
             return 'succ'
         return 'fail'
+    
+    def func_set_args(
+            self,
+            providers,
+            execution_threads,
+            keep_fps,
+            keep_frames,
+            skip_audio,
+            temp_frame_format,
+            temp_frame_quality,
+            output_video_encoder,
+            output_video_quality
+            ):
+        if ','.join(providers)!=','.join(roop.globals.execution_providers):
+            if enhancer is not None:
+                enhancer.clear_face_enhancer()
+            roop.face_analyser.clear_face_analyser()
+            face_swapper.clear_face_swapper()
+            roop.globals.execution_providers = providers
+
+        if execution_threads>0:
+            roop.globals.execution_threads=execution_threads
+        roop.globals.keep_fps=keep_fps
+        roop.globals.keep_frames=keep_frames
+        roop.globals.skip_audio=skip_audio
+        roop.globals.temp_frame_format=temp_frame_format
+        roop.globals.temp_frame_quality=temp_frame_quality
+        roop.globals.output_video_encoder=output_video_encoder
+        roop.globals.output_video_quality=output_video_quality
+    
+    def func_get_available_providers(self):
+        return onnxruntime.get_available_providers()
 
     def func_swap_image(self,source_face_infos,target_face_infos,target_img_path,output_file):
+        print("Checking nsfw...")
+        if roop.predictor.predict_image(target_img_path):
+            return 'nsfw'
         source_cache={}
-        target_img = cv2imread(target_img_path)
-        all_target_faces=roop.face_analyser.get_many_faces(cv2.cvtColor(target_img,cv2.COLOR_RGB2BGR))
+        target_img = cv2imread2rgb(target_img_path)
+        all_target_faces=roop.face_analyser.get_many_faces(rgb2bgr(target_img))
         result=target_img
         if all_target_faces is not None and len(all_target_faces)>0:
             min_len=min(len(source_face_infos),len(target_face_infos))
@@ -269,7 +352,7 @@ class MyHTTPHandler(BaseHTTPRequestHandler):
                     if info["file"] in source_cache:
                         source_img_faces=source_cache[info["file"]]
                     else:
-                        source_img=cv2.cvtColor(cv2imread(info["file"]),cv2.COLOR_RGB2BGR)
+                        source_img=rgb2bgr(cv2imread2rgb(info["file"]))
                         source_img_faces=roop.face_analyser.get_many_faces(source_img)
                         source_cache[info["file"]]=source_img_faces
                     source_face=source_img_faces[info["face_index"]]
@@ -283,7 +366,7 @@ class MyHTTPHandler(BaseHTTPRequestHandler):
                     if info["enhance"]==True:
                         target_face=all_target_faces[info["face_index"]]
                         result=face_enhance(target_face,result)
-        cv2.imwrite(output_file, result)
+        cv2imwrite(output_file, result)
         print("\n\nImage saved as:", output_file, "\n\n")
         return 'succ'
 
@@ -310,7 +393,7 @@ if __name__ == '__main__':
     else:
         port=53499
     roop.core.parse_args()
-    roop.globals.log_level='info'
+    #roop.globals.log_level='info'
     if not roop.core.pre_check():
         exit(1)
     if not face_swapper.pre_check():
